requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-huggingface
faiss-cpu
pypdf
python-dotenv
duckduckgo-search
sentence-transformers